Make compatible with Sidekiq 6.5.0
Sidekiq 6.5.0 refactored many internal APIs, one of which was
the Sidekiq::Util class. That class provided the #identity method
which was used to build a string for the processor label. The
functionality of that method was moved to a new Sidekiq::Component
module. This change uses the new module's #identity method.